Serapando os métodos da classe LeituraGravacaoArquivo
