--- dados.csv
@@ -1,8 +1,0 @@
-Maria dos Santos;01234567890;20000228;3500.00;
-João;00011122234;20180301;SOBRINHO;
-Joana;09876543201;20170101;FILHO;
-
-Raquel;01256567450;20000628;9000.00;
-Bruno;00011122235;20190301;FILHO;
-
-Rafael;01245257450;20100628;6900.00;
