--- folhaPagamento.csv
@@ -1,4 +1,0 @@
-Roberta dos Santos;52104959063;313,40;96,55;3090,05
-Rodolfo;83953994050;951,62;1317,30;6731,08
-Ricardo;39439074060;775,58;788,22;5336,20
-Ricardo;39439074060;775,58;788,22;5336,20
