--- folha_pagamento.csv
@@ -1,3 +1,0 @@
-Luiz dos Santos;10004959000;313,40;41,57;3145,03
-Roberto;08053994000;951,62;1265,99;6782,39
-Lorena;30039074055;775,58;788,22;5336,20
