+++ src/org/serratec/main/Main.java
@@ -4,17 +4,27 @@ import java.sql.SQLException;
 import java.util.List;
 
 import org.serratec.excecoes.DependenteException;
-import org.serratec.projeto.LeituraGravacaoArquivo;
+import org.serratec.projeto.Funcionario;
+import org.serratec.projeto.Gerar;
+import org.serratec.projeto.LeituraArquivo;
+import org.serratec.projeto.Processar;
+
 
 public class Main {
 
 	public static void main(String[] args) throws DependenteException, SQLException { 
-		LeituraGravacaoArquivo a = new LeituraGravacaoArquivo();
-		List<String[]> dadosPorLinha = a.lerArquivo();
+		LeituraArquivo leitor = new LeituraArquivo();
+		List<String[]> dadosPorLinha = leitor.lerArquivo();
+		
 		if (!dadosPorLinha.isEmpty()) {
-			a.processarArquivo(dadosPorLinha);
-			a.gerarCPFDuplicados();
-			a.gerarFolha();
+			Processar processador = new Processar();
+			processador.processarArquivo(dadosPorLinha);
+			List<Funcionario> funcionarios = processador.getFuncionarios();
+			List<Funcionario> funcionariosDuplicados = processador.getFuncionariosComCpfDuplicado();
+			
+			Gerar gerador = new Gerar();
+			gerador.gerarCPFDuplicados(funcionariosDuplicados);
+			gerador.gerarFolha(funcionarios);
 		}
 	}
 }
